Move TranslateExternsionConverter to its own source file
